pyproject.toml
[project]
name = "power-tracker"
version = "0.1.0"
description = "System power usage tracker using lm-sensors and PostgreSQL"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "flask>=3.0",
    "pika>=1.3",
]

[project.scripts]
power-tracker = "power_tracker.main:main"
power-tracker-client = "power_tracker.client:run_client"
power-tracker-consumer = "power_tracker.consumer:run_consumer"

[project.optional-dependencies]
dev = ["pytest>=9.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/power_tracker"]
